chores: refactor validateRequest test and added new
